client/requirements.txt
fastapi
uvicorn[standard]
psutil
websockets
pydantic
python-dotenv
psycopg2-binary
